reorg some functions
